requirements.txt
azure-ai-projects
azure-identity
azure-ai-inference
dotenv
flask
selenium
beautifulsoup4
pdfkit
python-dotenv
psutil
